+++ gdsc-smlm/src/main/java/uk/ac/sussex/gdsc/smlm/engine/DataEstimator.java
@@ -25,12 +25,12 @@
 package uk.ac.sussex.gdsc.smlm.engine;
 
 import java.util.Objects;
-import org.apache.commons.math3.stat.descriptive.rank.Percentile;
-import org.apache.commons.math3.stat.ranking.NaNStrategy;
+import org.apache.commons.statistics.descriptive.Quantile;
 import uk.ac.sussex.gdsc.core.threshold.AutoThreshold;
 import uk.ac.sussex.gdsc.core.threshold.AutoThreshold.Method;
 import uk.ac.sussex.gdsc.core.threshold.FloatHistogram;
 import uk.ac.sussex.gdsc.core.threshold.Histogram;
+import uk.ac.sussex.gdsc.core.utils.MathUtils;
 import uk.ac.sussex.gdsc.core.utils.NoiseEstimator;
 import uk.ac.sussex.gdsc.core.utils.Statistics;
 
@@ -186,21 +186,13 @@ public class DataEstimator {
    */
   public float getPercentile(double percentile) {
     // Check the input
-    if (percentile <= 0) {
-      percentile = Double.MIN_NORMAL;
-    }
-    if (percentile > 100) {
-      percentile = 100;
-    }
-
-    // The data should not have NaN so we ignore them for speed.
-    final Percentile p = new Percentile(percentile).withNaNStrategy(NaNStrategy.FIXED);
+    final double p = MathUtils.clip(0, 1, percentile / 100.0);
     final int size = width * height;
     final double[] values = new double[size];
     for (int i = 0; i < size; i++) {
       values[i] = data[i];
     }
-    return (float) p.evaluate(values);
+    return (float) Quantile.withDefaults().evaluate(values, p);
   }
 
   /**